REFACTOR | Dependencies, Interfaces and Extensions
Reworked: Dependencies, added Generic Interface to allow for easier creation of dependencies. For dependencies to work they must implement an Instance Property, before this was not possible unless the Dependency` type was used.

Reworked: IApiHandlerFactory, removed register handler as this is not need in this case.

Reworked GetInstance method for Dependency extension, it now used the new generic interface which allows better extensibility