ADDED  | SereneApi.Core and Handlers
As this will be a major rewrite of the code this is going to come as a new release with some signature changes.

SereneApi.Abstractions has been renamed to SereneApi.Core as this better describes the libraries role in the project.

SereneApi has been renamed to SereneApi.Handlers.Rest as Rest and Soap support will be in different packages to maintain SOC.

So far Configuration has been drastically changed through the use of ConfigurationProviders. During startup the ConfigurationManager will attempt to load all providers, and as each ApiHandler is registered the provider will be retrieved using the ConfigurationProviderAttribute. This means you're also able to use this in your project for finer control over your ApiHandlers without cluttering your startup.cs file.

This is still in the early stages of development so it is subject to change but so far early testing and semantics looks promising