Removed unused classes
Two classes in the DependencyInjection namespace were deleted as they were not being used. The first class was an internal sealed class that implemented IApiResourceCollection, and the second one was a static extension class for IServiceCollection. Both of these classes had methods that were not implemented or used elsewhere in the codebase.